Added block properties in the form of an unordered map of block types

+++ src/texture/TextureAtlas.h
@@ -7,13 +7,13 @@
 struct TextureAtlas
 {
     public:
-        static std::array<Vector2f, 4> getTexture(Vector2f p_pos)
+        static std::array<Vector2f, 4> getTexture(Vector2i p_pos)
         {
-            float x  = p_pos.x / (atlasSize / textureSize);
-            float sx = (p_pos.x + 1) / (atlasSize / textureSize);
+            float x  = (float)p_pos.x / (atlasSize / textureSize);
+            float sx = ((float)p_pos.x + 1) / (atlasSize / textureSize);
 
-            float y  = 1 - (p_pos.y / (atlasSize / textureSize));
-            float sy = 1 - ((p_pos.y + 1) / (atlasSize / textureSize));
+            float y  = 1 - ((float)p_pos.y / (atlasSize / textureSize));
+            float sy = 1 - (((float)p_pos.y + 1) / (atlasSize / textureSize));
 
             std::array<Vector2f, 4> texCoords = {
                 Vector2f(x,  y),
